Display changes for web interface
This in-development update allows for the now playing table to show what is upcoming for the next few hours on each channel. There are still a few bugs to work out, including those with ordering and the inability to show anything upcoming after midnight, until midnight has passed, and a few features to add (for example, skipping forward or back in the schedule).